refactor(Money) : 상수 자료형 int -> long 변환

+++ src/main/java/lotto/domain/Money.java
@@ -1,7 +1,7 @@
 package lotto.domain;
 
 public class Money {
-    private static final int MONEY_OF_ONE_TICKET = 1000;
+    private static final long MONEY_OF_ONE_TICKET = 1000;
 
     private final long value;
